Change resource directory, more work on gradients, and start on different types of ItemEntry for slightly better memory usage.

+++ src/main/java/mortvana/melteddashboard/client/texture/GradientNode.java
@@ -4,7 +4,7 @@ public class GradientNode {
 	public int color;
 	public float shade;
 
-	public GradientNode(int color, int shade) {
+	public GradientNode(int color, float shade) {
 		this.color = color;
 		this.shade = shade;
 	}
